+++ src/rex.core/NEWS.rst
@@ -9,6 +9,7 @@
 ==================
 
 * Added ``LatentRex`` class (Kyrylo Simonov).
+* Added NIH acknowledgement (Clark Evans).
 
 
 1.0.0 (2013-10-11)