+++ src/rex.core/NEWS.rst
@@ -8,6 +8,7 @@
 1.1.0 (2013-XX-XX)
 ==================
 
+* ``RecordVal`` supports field names that coincide with Python keywords.
 * Added ``LatentRex`` class (Kyrylo Simonov).
 * Added NIH acknowledgement (Clark Evans).